users/serializers.py
from rest_framework import serializers

from users.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        exclude =['password','is_admin','is_staff','created_at']